// Checkout/src/app/Core/Models/call-offer-dto.ts
export class CallOfferDto {
     callerId: string ='';
     calleeId: string ='';
     callType: string ='';
     id: string = '';
     callState: boolean = false;
     startCall:boolean=false;
     callerPic : string ='';
     calleePic  : string ='';
     callerName : string ='';
     calleeName : string ='';
}